backend/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
gunicorn>=20.0
google-cloud-aiplatform>=1.25.0
langchain-google-vertexai>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.20
SQLAlchemy>=2.0 
pandas>=1.3
pydantic>=1.8